+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.48.2] - 2021-07-31
+### Fixed
+- Dependency error in `remote` commands
+- Remove incorrect `.gitignore` file generation
+
 ## [0.48.1] - 2021-07-26
 ### Fixed
 - `sql` command failing on OSX